BlockNodeWithCapacity removed
