+++ ru.bmstu.rk9.rao.ui/src/ru/bmstu/rk9/rao/ui/process/blocks/BlockNode.java
@@ -224,7 +224,7 @@ public abstract class BlockNode extends Node {
 		return getName() + "." + dockName;
 	}
 
-	public final void validateConnections(IResource file, int sourceConnections, int targetConnections)
+	protected final void validateConnections(IResource file, int sourceConnections, int targetConnections)
 			throws CoreException {
 		if (getSourceConnections().size() == sourceConnections && getTargetConnections().size() == targetConnections)
 			return;
